periodically ping and properly reconnect upon disconnects
This commit uses AnyEvent::XMPP::Ext::Ping (part of the AnyEvent-XMPP
distribution) to send a jabber ping (XEP-0199) to the server every 60
seconds. If it doesn’t receive an answer for another 60 seconds, it
disconnects, which triggers a reconnect.

I tested this using an iptables DROP rule for the destination jabber
server and it worked so far. Let’s see how well it holds up in practice.